REF: extract clean_diff into separate module
